# [Go] Optimise HLL Update operation on arrays