--- a/hll/hll_8array.go
+++ b/hll/hll_8array.go
@@ -24,7 +24,6 @@ import (
 // hll8ArrayImpl Uses 6 bits per slot in a packed byte array
 type hll8ArrayImpl struct {
 	hllArrayImpl
-	configKMask int
 }
 
 type hll8Iterator struct {
@@ -68,19 +67,14 @@ func (h *hll8ArrayImpl) ToUpdatableSlice() ([]byte, error) {
 func newHll8Array(lgConfigK int) hllArray {
 	return &hll8ArrayImpl{
 		hllArrayImpl: hllArrayImpl{
-			hllSketchConfig: hllSketchConfig{
-				lgConfigK:  lgConfigK,
-				tgtHllType: TgtHllTypeHll8,
-				curMode:    curModeHll,
-			},
-			curMin:      0,
-			numAtCurMin: 1 << lgConfigK,
-			hipAccum:    0,
-			kxq0:        float64(uint64(1 << lgConfigK)),
-			kxq1:        0,
-			hllByteArr:  make([]byte, 1<<lgConfigK),
-			auxStart:    hllByteArrStart + 1<<(lgConfigK-1),
-			configKMask: (1 << lgConfigK) - 1,
+			hllSketchConfig: newHllSketchConfig(lgConfigK, TgtHllTypeHll8, curModeHll),
+			curMin:          0,
+			numAtCurMin:     1 << lgConfigK,
+			hipAccum:        0,
+			kxq0:            float64(uint64(1 << lgConfigK)),
+			kxq1:            0,
+			hllByteArr:      make([]byte, 1<<lgConfigK),
+			auxStart:        hllByteArrStart + 1<<(lgConfigK-1),
 		},
 	}
 }
@@ -124,7 +118,7 @@ func convertToHll8(srcAbsHllArr hllArray) (hllSketchStateI, error) {
 
 func (h *hll8ArrayImpl) couponUpdate(coupon int) (hllSketchStateI, error) {
 	newValue := coupon >> keyBits26
-	slotNo := coupon & h.configKMask
+	slotNo := coupon & h.slotNoMask
 	err := h.updateSlotWithKxQ(slotNo, newValue)
 	return h, err
 }