--- a/hll/hll_array.go
+++ b/hll/hll_array.go
@@ -63,8 +63,6 @@ type hllArrayImpl struct {
 
 	auxHashMap *auxHashMap
 	auxStart   int //used for direct HLL4
-
-	configKMask int // mask from lgConfigK
 }
 
 // newHllArray returns a new hllArray of the given lgConfigK and tgtHllType.