Precompute the slotMask once per HLL/config instead of per update

```
goos: darwin
goarch: arm64
pkg: github.com/apache/datasketches-go/hll
                                  │ old-all-bench.txt │         new-all-bench.txt          │
                                  │      sec/op       │   sec/op     vs base               │
HLLDataSketch/lgK4_HLL4_uint-10           18.67m ± 1%   18.07m ± 2%  -3.24% (p=0.002 n=12)
HLLDataSketch/lgK16_HLL4_uint-10          18.68m ± 1%   18.16m ± 1%  -2.81% (p=0.000 n=12)
HLLDataSketch/lgK21_HLL4_uint-10          20.35m ± 0%   20.00m ± 0%  -1.72% (p=0.000 n=12)
HLLDataSketch/lgK4_HLL6_uint-10           19.01m ± 1%   18.34m ± 1%  -3.52% (p=0.000 n=12)
HLLDataSketch/lgK16_HLL6_uint-10          19.24m ± 1%   18.38m ± 0%  -4.44% (p=0.000 n=12)
HLLDataSketch/lgK21_HLL6_uint-10          21.32m ± 1%   20.13m ± 1%  -5.59% (p=0.000 n=12)
HLLDataSketch/lgK4_HLL8_uint-10           17.19m ± 1%   16.83m ± 0%  -2.13% (p=0.000 n=12)
HLLDataSketch/lgK16_HLL8_uint-10          17.28m ± 0%   16.89m ± 0%  -2.25% (p=0.000 n=12)
HLLDataSketch/lgK21_HLL8_uint-10          18.68m ± 0%   18.32m ± 0%  -1.94% (p=0.002 n=12)
HLLDataSketch/lgK4_HLL4_slice-10          19.29m ± 0%   18.83m ± 0%  -2.39% (p=0.000 n=12)
HLLDataSketch/lgK16_HLL4_slice-10         19.50m ± 1%   18.89m ± 0%  -3.13% (p=0.000 n=12)
HLLDataSketch/lgK21_HLL4_slice-10         21.15m ± 1%   20.46m ± 0%  -3.27% (p=0.000 n=12)
HLLDataSketch/lgK4_HLL6_slice-10          19.71m ± 0%   19.15m ± 0%  -2.83% (p=0.000 n=12)
HLLDataSketch/lgK16_HLL6_slice-10         19.85m ± 1%   19.24m ± 0%  -3.07% (p=0.000 n=12)
HLLDataSketch/lgK21_HLL6_slice-10         21.14m ± 0%   20.58m ± 0%  -2.69% (p=0.000 n=12)
HLLDataSketch/lgK4_HLL8_slice-10          18.07m ± 0%   17.67m ± 0%  -2.21% (p=0.000 n=12)
HLLDataSketch/lgK16_HLL8_slice-10         18.14m ± 1%   17.74m ± 0%  -2.23% (p=0.000 n=12)
HLLDataSketch/lgK21_HLL8_slice-10         19.58m ± 1%   19.14m ± 1%  -2.23% (p=0.000 n=12)
HLLDataSketch/lgK4_HLL8_union-10          23.42n ± 0%   23.34n ± 1%  -0.36% (p=0.016 n=12)
HLLDataSketch/lgK16_HLL8_union-10         23.40n ± 0%   23.34n ± 0%  -0.26% (p=0.008 n=12)
HLLDataSketch/lgK21_HLL8_union-10         23.44n ± 1%   23.36n ± 0%  -0.32% (p=0.009 n=12)
geomean                                   2.749m        2.680m       -2.51%
```